Created: The base object class of things that have ID numbers
For any object that has an ID number (if there are ID's that have
characters in them, I will convert to string) this will be used. It
automatically adds __slots__ (to save memory if many are to be cached)
and it allows them to be compared with equal and not equals.