Created: Messge object
Many things are NotImplemented in this object. Lots of TODO notices.
Some things in the API I do not know what they do. May need to
investigate more, or it also may forever be a mystery. Will probably
make an object for the content of the message because looking at the API
it is actaully complicated how it is setup. Reactions will have an
object, but that has always been my plan.