Implemented the idea for how lobbies will be added to the message objects. It will return the Lobby object that is passed with the kwargs when initialized. Layout of how it will work:
1. Bot connects to community

2. Bot caches the community, with a list of the lobbies in the cached
community object.

3. Message is received.

4. Using the 'lobby_id' attribute of the message's JSON object, search
through either all of the communities and their lobbies, or find a way
to only look through the community is was sent through's lobbies. I do
not think I will make a separate lobby cache because they will already
be cached with the community. Can just look through the communities
with a For loop.

There is also a minor formatting change.