Created: Presence object
This object represents the presences that the Spectrum API spits out
about members. Not too hard, just a few attributes. The harder part
will be the NotImplemented that use it, there seem to be quite a few.
I also put the object in the libraries __init__ already. Learned a
little more about autodoc since making the lists of the values the
attributes can be was a little bit of a challenge for them to compile
with no errors. I will add the autodoc lists to other attributes in
other objects that have small limited amount of values.