Created: Member attribute member.presence
The attribute for the presence of a member has been added. It allows to
be added from the cache when the member object is created. I do not
know if that will be useful, but is there in case. If this in not
useful at all I will remove it, maybe. If I remember to. As I do often
in these early commits I will outline how I think the netcode will work
later.

1. Connected to lobby
2. From the REST API get all presences from the members in the initial
presence POST.
3. Create Member objects and cache them
4. Presence are part of members so they are cached with them
5. If a presence updates run a hidden method that updates the presence

I may write that presence updating method right now. Or maybe when I
get to writing networking. I should probably do it now. In fact, all
objects should have a hidden update method now that I think about it.
I do not know what parts of all the objects can update on the fly, so
I may rewrite all of the initialization code for the objects. RIP in
peace me.