Created: entities file and Emoji entity object
I created a single file for all the entities because there are several
and they are not that long. It also allows me to organize them if more
are added. I will also add the mention and link entities soon.

The emoji object is a little weird. It will reside in a list in the
Block objects. It will also gobble up the "entityRanges" JSON object,
because I cannot think of any other way to do it. I think this is the
best way. Also I do not think I will ever be able to convert between
Unicode and the emoji objects. There is a lot of Javascript in Spectrum
that goes on to do this, but manually decoding all of it would be pain
in the butt. So I will not. If there is another way, let me know.