Created: Spectrum Mention object
This is an object that represents mentions for the future (soon^tm)
object that represents Spectrum blocks. Not too interesting except that
the member attribute may default to None. This happens will if the
member is not in the cache of the client when this mention is received.

There are also a few random formatting changes for docstrings