Removed: Assumption that things always work perfectly
I assumed that if the member attribute of the mention object isn't None
that it has to be a Member object. Who knows what will happen once
networking starts. So I added a check to make sure that it is an
instance of the Member object before I reference the name attribute of
the member attribute.