Created: Spectrum Block object
This object was probably the hardest to make overall. Lots of lead up
to do. I had to create entities and the style_ranges in order to make
it. In the end I think I am happy with how it should turn out. I may
add a hash(x) ability because this object is able to due to the ID
not being an int. The way in which the entities are solved to their
respective objects is a little messy, and I do not like it too much,
but I do not like if -> elif chains much anyways. I am getting so close
to being able to start learning how to network everything. Hopefully
before I move I will be able to start. I also added the Block object to
the documentation already, proactive.