Changed: Community lobbies attribute returns iterator
Before it returned the type dict_values, which I forgot about. I just
assumed it return a list or an iterator/generator, but it does not. So
I consulted the experts of /r/starcitizen discord, and they recommended
returning an iterator instead of a list. Thank you @jason-weirather for
your contribution to this commit.